examples: Remove pr2_passive_simulation
Absent an owner for this incomplete demo based on RigidBodyTree, we are
removing it until someone wishest to resurrect it for MultibodyPlant.

We keep the model files intact, because they are intalled, used elsewhere
in the tree, and porting the unit test is easy.

The unit test shows that the MbP URDF parser is more strict; so we also
repair the URDF to be valid.