examples: Remove double_pendulum (#12725)
* examples: Remove double_pendulum

This demonstration was a stepping stone showing how to use libsdformat
to parse SDFs into a simulation.  In the years since this was added,
we've introduced a multibody::Parser targeting MultibodyPlant that
uses libsdformat for all of its *.sdf parsing.

Since this demo is one of the sole reamining uses of RigidBodyTree,
it seems time to retire it and point users to multibody::Parser
instead.