[geometry] SceneGraph given a new query: hydroelastic with (point pair) fallback (#12699)
Introduces new query that attempts to compute hydroelastic contact surfaces.
However, if, for whatever reason, a geometry pair doesn't suport
hydroelastic contact, it computes the penetration as point pair for that
pair.

The refactor of callbacks led to a significant respelling of the unit test;
we went from testing one function to three with similar requirements on
the underlying infrastructure.

Updates the simple vis contact to exercise this new query.