Add plane-tet intersection
This is the atomic operation to support rigid half space vs soft volume
mesh intersection. It robustly intersects a single tet with a plane in
such a way that no redundant edge-plane intersection tests are performed
and (as a by-product) doesn't introduce duplicate vertices in the resulting
contact surface mesh.