Changing how polygons get split for contact surfaces
Every triangle/polygon found in ContactSurfaces needs to be split into
a triangle fan. A new vertex is calculated on the interior of the face
and the fan is defined around that vertex.

Previously, the centroid vertex was the average vertex position. This,
however is not stable. As a quad becomes a triangle, it causes the vertex
location to jump. Instead of the average vertex position, we use the
polygon's center of mass. This changes continuously, regardless of the
exact distribution of vertices.