Clean up plane.h
Plane used to be a local computational half space. In the past it was
renamed to Plane. However, it had a number of vestiges remaining of the
half space origin. This cleans up the documentation and the API (and its
usages) to provide a more coherent description and implementation of a
Plane, proper.