Clean up plane.h
Plane was originally internal::HalfSpace. It had previously pulled out to
be used more widely. To avoid collision with the geometry::HalfSpace name
it was renamed Plane. However, this led to confusion due to the fact that
it was reporting a *signed distance*.

The resolution: keep the definition of internal::Plane but give it "height"
semantics (indicating distance "above" or "below" the plane). Also include
a "posed half frame" (defined by its boundary plane). This will allow us
to pose a half space or plane in an arbitrary frame, but far more clearly
communicate the semantics of the intersections (planes have no volume but
half spaces have infinite volume).

This performs the renaming, updates usages and documentation.