[parsing] Add unit test to cover Parser's automatic search for packages (#12871)
The Parser class will attempt to locate packages for every sdf/urdf
passed to one of its Add*FromFile() methods. However, it will only
work if the model file is located in the Drake tree. This adds
a test that confirm this behavior.

Incidentally, it does some minor clean up on the package_map
documentation.

This can be omitted from the change log in future releases.