framework: Update internal for ValidateContext rename (#12891)
In 54717e9aa3dce593a2fd3277de3c7575351834e8 we added a fast new method
to SystemBase for validating contexts, and deprecated the old ones.  However,
one old method was also exposed via internal::SystemMessageInterface but
the commit did not update that API.  We do that here now.

This turns up a bug in unit test code where a MultibodyPlant was losing
its system_id during transmogrification, because we were accidentally
transmogrifying it implicitly(!) while passing it into a function,
because MultibodyPlant's scalar-converting copy constructor was not
marked explicit.

This is very bad, so we also add explicit to MultibodyPlant now.
This is a breaking change but well worth it.

We also grep the code for any other missing explicit on such
constructors, fixing just the one on GeometryState.