Updating contact surface utility to handle degeneracies
This was born of a need to handle degenerate polygons (polygons with
edges of length zero). It invited a revisiting of what the invariants
on the parameters should be.

So, we handle degeneracies and provide a debug-build analysis to help
protect the function from bad values. All of this is now under test.

 1. Add DRAKE_ASSERT_VOID to validate inputs
    1. Confirm that the normal vector has meanginful length.
    2. Confirm that it is sufficiently perpendicular to the plane.
    3. Confirm polygon is planar.
 2. Document these prerequisites and indication that they will
    (somewhat be tested in debug builds).
 3. Better handle degenerate polygons -- i.e., if it has zero area,
    don't return a centroid that is all NaNs.