analysis: Add default_scalars annotations where still absent (#12908)
The current design intent of the simulator and its integrators is to
support the default nonsymbolic scalars only.  Previous commits had
added annotations to that effect in some places, but missed others.
This commit ensures that the simulator, all of its integrators, and
its supporting classes are all uniformly annotated to support only
the default nonsymbolic scalars.

This commit only change documentation and adds new extern template
declarations; it does not yet remove the ability to use non-standard
scalars, though we plan to do so in the future.  In the future, we
may also extend support to add symbolic scalars.

Relatedly, whitespace style choices related to the default scalar
boilerplate are also unified here to be consistent across all files.