Move MBP::HydroelasticContactInfoAndBodySpatialForces into internal
MBP has a nested class that is stored in a cache entry. The hash
logic associated with the drake::Value type doesn't work well with
nested classes of templated classes (it introduces a "namespace"
with a template parameter). This clutters up the console with
warnings that don't help the end user at all. SO, for now, we'll
pull the nested class out and put it in an internal namespace.