[render] Modify vtk cube texture coordinate generation (#12972)
* Modify vtk cube texture coordinate generation

The default vtk cube source produces texture coordinates based on the
size *and* position of the cube in the geometry frame. More or less, the
vertex position becomes a uv coordinate (projected onto one of six planes).
This has several undesirable properties:

  1. Small boxes will only be mapped with small portions of the texture.
  2. Large boxes will have the texture tiled multiple times.

This introduces a new cube geometry that is equivalent to the vtk cube
source, except it defaults to the texture being stretched over each face,
with the possibility of tiling it via a (currently hidden) uv_scale
property. It also removes all mention of the vtkCubeSource.