Fix math::jacobian() segfault when evaluated on constant AutoDiffXd function
When math::jacobian() is ran on an f(x) with a return scalar type of AutoDiffXd, it will segfault if f(x) does not depend on x. This is because the derivative of the result has size 0, while math::jacobian() tries to index into the f(x) result derivative vector to set its final vector.

This commit:
1. Adds a test into jacobian_test.cc to reproduce this bug, which creates a constant f(x) = 0 of scalar type AutoDiffXd for it to differentiate.
2. Fixes this issue by having the inner for loop stop early, and setting the rest of the derivative vector to 0.