trajectories: construct HermitianDenseOutput from PiecewisePolynomial (#13006)
now with support for default scalars.
(I should have done this version in #12957, but didn't realize it until I was able to push forward the default scalar support for PiecewisePolynomail).  This technically changes the API that I introduced last week; i think it's safe to assume that nobody was using it yet.