integrators: replace DenseOutput with PiecewiseTrajectory (#13008)
* integrators: replace DenseOutput with PiecewisePolynomial

This breaks an API (without deprecation) by changing the return type of `IntegratorBase::StopDenseIntegration` and `IntegratorBase::get_mutable_dense_output` to be a `PiecewisePolynomial`.  After dicussion with @sammy-tri, we decided that this API was officially deep and obscure to warrant this change.  `InitialValueProblem` was the only known consumer of this method (in general, we do not have workflows that would work with the `Integrator` directly, and these APIs were not exposed through `Simulator`.