Add alternate Sphere volume mesh generation
Rather than creating a sphere by sub-dividing *all* edges of the
previous level of refinement, we refine only those edges on the
surface of the sphere.

The resultant mesh has a single interior vertex (at the sphere
origin) and all of the resolution increases curvature. In contrast
to the interior vertices which increases the number of tets by a
factor of 8 each time, this only increases it by a factor of 4.

It retains an option to tessellate on the interior as well, but
no longer defaults to that.

This also provides a mechanism through proximity properties to
toggle what type of sphere tessellation is used.
An optional parameter (kHydroGroup, "sparse_mesh") has
been added which will control whether the sphere gets tessellated
on the interior or not (defaults to not).

This cascaded to a number of other pieces of code that made use of
the sphere. The ellipsoid representation as well as the tests.