Introduce the MeshPainterSystem
The MeshPainterSystem uses contact between two geometries to
paint onto a texture. One geometry is the "canvas" geometry
and has a texture map associated with it. The other geometry
is the painter object. When the painter and canvas geometries
make contact, the region of contact is drawn onto the canvas
object's texture map. The system ouputs the *accumulated*
texture on the output port.

This is largely a prototype and has the following issues:

 - Contact patch computation needs optimization.
 - Limited support for canvas geometries
   - only meshes are supported (and they must come from OBJ
     files that have "valid" texture coordinates).
   - The definition of "valid" texture coordinates is very
     strict (see documentation in mesh_painter_system.h).
 - Limited support for painter geometries
   - Must be geometry we can currently create volume meshes
     for (box, cylinder, ellipsoid, sphere)