[doc] Remove the incorrect drakeURDF.[xpr|xsd] (#13326)
The schema defined in drakeURDF.xsd is significantly out of date. We don't
have the time or resources to bring it up to date. For now, we'll just
remove it so as to avoid the suggestion that it is reliable. It can always
be resurrected from git when it becomes a priority.

This also removes references to the file.