Resolve signed-unsigned comparison issues in TypeSafeIndex
When comparing an unsigned integral value with a TypeSafeIndex, we
needed to account for the possibility that the unsigned integral value
could contain larger values than the TypeSafeIndex. This was done by
comparing the unsigned integer with the maximum int (the largest type safe
index value). This is intrinsically a signed-unsigned comparision which is
problematic.

Instead, we cast the maximum int to the unsigned integral type. If the
unsigned type is larger (more bits), the relationship will be fully
preserved and unsigned values that are larger than can be represented will
be detected.

If the unsigned type has fewer bits, the maximum int (01111.1111) will be
truncated to a number consisting of all 1s, the maximum unsigned value of
that number of bits. We are guaranteed that the unsigned _value_ must
be less than or equal to that truncated value which means its less than or
equal to the full maximum index.

In other words:

value <= static_cast<U>(kMaxIndex) is always *true* if
    sizeof(U) < sizeof(int).