Fix symbolic::Expression::Expand()
In `ExpandMultiplication(e1, e2)`, it returns `e1 * e2` if both `e1` and `e2` are non-additive expressions. However, if at least one of them is a division expression, it is possible that `e1 * e2` is not an expanded expression.

For example, let `e1 = (x + 1) / y` and `e2 = (x + 3) / z`, `e1 * e2` becomes `(x + 1) * (x + 3) / yz` whose numerator is not expanded yet.

This commit fixes this problem by handling those cases.