Step doubling for Implicit Euler Error Estimation (#13224)
* Step Doubling for Implicit Euler Error Estimation

This PR implements Implicit Euler with step doubling for error estimation,
similar to how VelocityImplicitEulerIntegrator implements it.

The major change from VIE's implementation is that Implicit Euler takes advantage of marking a Jacobian as "fresh" after a failed step, and never recomputing the Jacobian as it shrinks step sizes. The problem is, now, after it succeeds on the large step and the first small step, there is a rare (but nonzero) probability that it fails on the second small step, computing the Jacobian at a state that is not xt0. (In a stiff simulation of about 2000 steps with about 2000-4000 convergence-based failures, this scenario will occur maybe about ~20 times.) Based on some thought (and confirmed by experimentation), I've proposed a set of small changes that captures the performance gains of marking the Jacobian as "fresh" without triggering too much extra overhead. Note that this is much simpler than the first revision of the PR (I've removed a lot of caching logic which only improves performance by 2%, and replaced it with a TODO, as Sherm suggested):

1.    After the Jacobian is computed, it is immediately marked as "fresh" (set_jacobian_is_fresh()).
2.    Before starting the second step, we mark the Jacobian as "not fresh" (set_jacobian_is_fresh(false)). This way MaybeFreshenMatrices will recompute the Jacobian if the second small step is needed.
3.    If the second small step actually fails, then we set a flag (failed_jacobian_is_from_second_small_step_) to indicate that at the beginning of the next step, the Jacobian should not be marked "fresh", even though ImplicitIntegrator<T>::DoStep() marks it as fresh.

This set of changes incurs minimal expenses during successful steps (only setting booleans), and it only forces the Jacobian to be recomputed again if the second small step actually fails.