RenderEngineGl gets ability to "show_window" (#13626)
* RenderEngineGl gets ability to "show_window"

The RenderEngine API has the ability to request showing the
display for color and label images. Up to this point, RenderEngineGl
hasn't supported that. This adds the infrastructure necessary
to display the images.

Currently only exercised in test. Its inclusion now facilitates
the development of label and color rendering down the road. It's
nice to be able to see the images.

(Incidentally changes opengel_context.h documentation to /* from /**
because it is in internal:: namespace.)

Co-authored-by: Belinda Teh <belzteh@gmail.com>