Eliminate (group, property) for just single string: name
The (group, property) paradigm was speculative. In the time the properties
have been extant, it has been shown that there's not much value there.
So, instead, we'll simply use single strings and rely on the user to add
prefixes/suffixes as they like to distinguish otherwise similar names.

The old API is largely expressed in terms of the new API. There are some
*very* subtle changes, however. We no longer have a *special* exception for
invalid groups. The tests have been modified accordingly.

The tests have been modified with an eye looking towards deprecating the
old (string, string) property API. The tests in their original state have
been copied and renamed GeometryPropertiesOld and the original tests have
been modified to use the new test. Other than changing which variant of
the API is being called, and the very specific group-related artifacts,
all tests are identical.

TODO:
 - update perma link in pydrake_doxygen.h. Thist can't be done until this
   commit lands in master.