yaml: Add SubtractDefaults to YamlWriteArchive
This allows for emitting a minimal yaml representation when loaded with
retain_map_defaults set on the eventual reader.  This is the minimal API to
accomplish that; I intend to provide better usability shims for this in a
future commit.

Also, move EmitString definition into the cc file -- non-template method bodies
should generally live there for better compile times.