Change property string representation
Finally, this changes the convention of the string representation of the
property name. The old format was ('group', 'property'), however, that
can't be used conveniently in regular expressions. I.e., if I did this:

EXPECT_EQ("The property: ('a', 'b')",
          fmt::format("The property: {}", property_name));

the test would fail, even though the two strings are identical because the
second string would be interpreted as a *regular expression* and the
parentheses wouldn't be matched characters. However, a/b *does* match.