PlanarSceneGraphVisualizer: Store convex hulls of meshes (#13857)
* PlanarSceneGraphVisualizer: Store convex hulls of meshes

Prior to this commit, we stored all of the vertices of each mesh and then took
their convex hull after projecting them to the visualization plane every dt. Now
we only store the vertices of the 3D convex hull of each mesh. The 2D convex
hull must still be computed at each dt, but, for meshes with lots of interior
vertices, it will now be over a much smaller set of points.