Provide System::get_input_port() and get_output_port()
The cost of authoring a new system is still too high.  I was about to implement the default (no argument) `get_input_port()` and `get_output_port()` methods for my new system, but realized that the amount of trivial code throughout the codebase was immense.  (Rule of >>three?).

Not having these default also complicates the python bindings.  Users that wanted both methods had to provide overloads of the base system class for the derived class.  And if I had a nickel for every time a student/novice user asked about why `get_input_port(0)` worked for some systems, but not all systems that have >0 input ports...

This PR also proposes that we STOP the practice of deleting the default `get_input_port(int)` and `get_output_port(int)` methods when we offer the `get_input_port()`, etc method alternatives.  In this case, I will happily trade the compile-time check for the runtime check in order to imho considerably improve the consistency of the API and the user experience.  (I was going to defer this change to a second PR, but it turns out they are coupled, since the deletions as written were deleting both base class overloads and causing compilation errors.)