framework: Mark subvector and supervector final
This will make the h/cc split clearer in the future, as well as
allow us to improve the implementations and inheritance tree of
the VectorBase family.

This is a breaking change, in the unlikely event that anyone was
inheriting from these classes.

Also deprecate a worthless constructor of Subvector.