MeshcatVisualizer: Tweaks to support caching mesh geometry on the zmq server
Adds an argument to the constructor to avoid deleting the entire tree on every run.
Sets uuids deterministically based on the mesh file so that they can be recognized as identical on the server.
See https://github.com/rdeits/meshcat-python/pull/75

This makes a *dramatic* improvement in the workflow of using meshcat on colab (at least as we use it in drake), because we don't have to repeatedly download large meshfiles from repeated simulations.