AutoDiffScalar: move-aware rewrite of arithmetic
Relevant to #13985.

See this comment in the ticket for benchmark results:
https://github.com/RobotLocomotion/drake/issues/13985#issuecomment-684987211

Summary of changes

* add compiler-default move-aware constructor and assignment
* rewrite arithmetic operators to take advantage of move semantics
  * make foo= the primitive and rewrite binary operators in terms of those.
  * use pass-by-value and rvalue optimizations to remove some allocations.
    * See Effective C++ Item 41 for discussion.

It is possible that pass-by-value and rvalue optimizations leave some
efficiency gains still on the table. Such gains would have to be won by
very careful overloading of templated functions within a templated
class, while (somehow) avoiding introduction of ambiguous overloads. In
contrast, this implementation achieves significant gains and is still
fairly readable.