DifferentialIKSystem: Add (optional) discrete state to "warn once" (#14004)
* DifferentialIKSystem: Add (optional) discrete state to "warn once"

This significantly improves the robustness of the teleoperation demo on colab.  Currently, if a user drives the robot into singularity (and it is relatively common mistake), then the diff IK spews so many warnings that the sliders disappear (and so does the "stop simulation" button).  This quiets that behavior significantly.