[geometry] API to get a frame's geometries (#13999)
* API to get a frame's geometries

This introduces a new API. Given a frame id, the ids of the frame's
geometries can be returned. It can be filtered on a particular role.

This includes the C++ and python interface.

(In the process of adding tests to the new pydrake bindings, I noticed that
there were tests missing for previously existing python APIs. They were
added and the *order* of the existing tests were changed to match the
order of the bindings to ease future reading.)