Geometry world pose as up to date as possible after Registration (#13990)
* Geometry world pose as up to date as possible after Registration

Prior to this, the world pose of the geometry was ignored upon
registration. It was assumed that any query that depended on geometry
world pose would be preceded by a call that updates the frames' poses.

However, it is possible to
  - Update frame poses
  - Add a geometry
  - perform a query

In that case, the added geometry will not have the correct world pose.

This change stores a value for X_WG based on the assumption that X_WF is
meaningful. If the assumption is correct, X_WG will likewise be meaningful.
If the assumption is wrong, we are not harmed.

Furthermore, for the queries (e.g., illustration, perception, and proximity
to work, those values need to propagate downwards). Work for proximity and
perception has been done. The illustration role has no further copies of
the pose beyond that in GeometryState.