framework: Regularize VectorBase bounds checking (#14101)
In preparing to move some methods into cc files, it became clear that a good
code split between h and cc would be difficult given the current idioms for
bounds checking.

This commit adjusts operator[] access to not check bounds (it was already
documented not to check, but was doing so anyway) and unifies the exception
formatting into the VectorBase class.

This is a breaking change in the sense that the exception subclass being thrown
may be different now; the API promises are weakened to only promise to throw a
std::exception, not any specific kind of exception.  The unit tests are
weakened to reflect this.

This is breaking change in demoting BasicVector::DoGetAtIndex from protected to
private.

Also add a Doxygen comment to specify SetZero.

Also adjust #include statements to remove unused inclusions.