cassie_bench: Separate autodiff malloc counts from benchmark timing
Relevant to: #10991, #13902

I finally realized that LimitMalloc counting was contributing significant
overhead to autodiff benchmark timings, owing to necessary synchronization
primitives in that module. This patch separates the two measurements, to
clarify the things we want to focus on.

Notice that this change of measurement will require some revision of our
timings for older versions, to keep comparisons sensible. Reviewers should look
for updates to the tracking issue #13902.