Fix throwing-Get methods missing Has counterparts (#14176)
If a class exposes a `Foo& GetFoo(string)` method that throws if
there is no Foo of that name, then it must expose a corresponding
`HasFoo` method that allows a caller to know whether the getter
will throw.  Otherwise the no-catches standard makes safe use of
the getter impossible.

We've stomped most of these in the codebase but I just found these
two stragglers.