SceneGraph can store geometry data as State or Parameter
While this provides the ability to choose whether the giant geometry data
blob for SceneGraph is stored as Parameter or State, the default behavior
remains State to maintain backwards compatibility.

In support of this endeavor, the LeafSystem interface is extended to
declare an abstract Parameter by passing ownership of a pre-instantiated
object. This enables SceneGraph to hang on to a pointer and mutate it
post declaration.

Tests are extended accordingly.