Rewrite of MeshcatContactVisualizer
I looked more closely at the code because I hit a bug where the contact force IDs were incrementing unstably even though my scene was in steady state.  It caused flickering contact forces, and ground the performance in meshcat to a halt.

I found the current implementation to be unnecessarily complex and inefficient.  This PR cleans it up.

As a bonus, I've also added arrow heads to the cylinders, and changed the colors to match the contact forces in drake visualizer.  The overall result improves the visualization quite a bit.