Add stl2obj to deal with robot stl meshes (#14219)
* Add stl2obj to deal with robot stl meshes

Convert the pr2 example to use stl2obj.  Previously, we had both copies
of the meshes in git.  Now, we keep only the upstream stl meshes in git
and we create the obj meshes during the build.

As a side-effect, we now create and install kinect_mount.obj, whereas
previously it was only installed in *.stl form.