Add FemElement and FemState.
FemState stores the node-wise states in a FEM simulation as well as
ElementCache that contains state-dependent quantities that are
associated with the FEM elements. FemElement consumes FemState and
performs the element routine in a FEM solver. Currently, the only
element routine we support is calculating the residual.

Concrete instance of ElementCache (ElasticityElementCache) and concrete
instance of FemElement (FemElasticity) are added to illustrate how
these two classes work together.