Deprecate geometry data as State in SceneGraph
This changes the default behavior in SceneGraph; a default-constructed
SceneGraph stores its geometry data as a Parameter (previously it was
State).

This leaves a stop-gap constructor in place (marked as deprecated) to
explicitly request the data to be stored as State.

We also patch two sets of unit tests that assumed geometry data was
stored as State (now Parameter).