[geometry] Deprecate geometry data as State in SceneGraph (#14245)
This changes the default behavior in SceneGraph; a default-constructed
SceneGraph stores its geometry data as a Parameter (previously it was
State).

Having geometry data as state meant that every diagram with a populated
SceneGraph had abstract state -- it couldn't be a strictly continuous system.
This interferes with desirable analytical techniques. 
Furthermore, it's only benefit (editing geometry via event) proved to
be basically worthless and largely impossible.

This leaves a stop-gap constructor in place (marked as deprecated) to
explicitly request the data to be stored as State.

We also patch two sets of unit tests that assumed geometry data was
stored as State (now Parameter).