Clean up and deprecation of SceneGraphInspector/QueryObject bindings (#14280)
* Clean up and deprecation of SceneGraphInspector/QueryObject bindings

As per issue 14258, we want all public APIs that take Identifiers as
parameters to have names representing their id type (source_id and not
simply id). This handles what is termed "Category 3" APIs in that issue.
This includes deprecations of some APIs that have been renamed and some
deprecations on only the bindings.