meshcat: upgrade interface to use QueryObject
The deprecation strategy here is more complicated than I would like.

I've added a new input port, but left the old one there.  I branch in the load and draw method based on which is connected.  I've also made the scene_graph argument in the constructor optional.

For ConnectMeshcatVisualizer, the story is slightly more complicated; my rules are
1) None of the current use cases can cause errors (only warnings).
2) If the use case is already using the syntax we want in the long run, they should get the new functionality with no changes and no warnings.
That was difficult to achieve.  The problem is that I can't know which API to offer them until `load()`.

In the tests, I have copied all of the old tests with a "_deprecated_api"; the new versions are only a few characters different, but this should make the removal at deprecation time better, I hope.