Update RenderEngine python trampoline - omit soon-to-be-removed API
The RenderEngine API has a Render*Image(CameraProperties) API. The
CameraProperties classes are being supplanted by the RenderCamera API
(which gives support for a fully specified pinhole camera intrinsic model).
Rather than bind these soon-to-be-deprecated methods, we'll omit them to
ease the pain of deprecation.

We're assuming that since these binding have only existed for about two
weeks (as of this commit), a hard removal won't cause too many problems.