Change property string representation
This changes how a property is described in error messages and comments
to move away from the soon-to-be-deprecated (group, property) notation to
just a single name.

For now, the name is group/property (because that is what is produced when
using the (group, property) API). This is subject to change.