testTemporary print statements to help debug an error
This decorates the problematic line of python with print statements
bounding the invocation (creation of RgbdSensor) and its conclusion
(RgbdSesnsor constructor in C++ prints the names of the cameras). In
between it prints a breadcrumb trail of the creation of various instances
of RenderCameraCore.

Specifically, one instance is created and then serves as, first, the source
of another instance's *move* constructor and *then* as the source of a
copy constructor. This second instance is what persists, but was created
from an instance whose values had been moved -- therefore, it has an
empty string as a renderer name (leading to the CI error).

This is observed in the output string as a single pointer appearing three
times.

(using GCC and *not* clang)
bazel test --compilation_mode=dbg //bindings/pydrake:py/geometry_test

The failing test will include a fair amount of test. The output of interest
is bounded at the beginning by the string:

"Testing RenderEngine via RgbdSensor"

and at the end by

Color camera renderer:
Depth camera renderer: renderer