Remove all dependencies on bot_core messages
Exorcize use of all bot_core messages within Drake, with the sole
exception of the python definitions for drake_visualizer.

The lcmtypes_bot2_core external and installed message definitions
are deprecated and will be removed on 2020-08-01.

List of breaking changes:

Remove message definitions for:
- robotlocomotion.grasp_transition_state_t
- robotlocomotion.robot_plan_t
- robotlocomotion.robot_plan_w_keyframes_t
- robotlocomotion.robot_plan_with_supports_t
These message depend on bot_core.