workspace: Avoid using whitespace in pkg-config rpaths
As of bazel 4.0, the user_link_flags changes from a list to a depset.

Therefore, when we used -Wl,-rpath /foo/bar as our spelling, the first
instance survived but a second instance would skips the -Wl,-rpath part
but keep the /quux/bar part, leaving a stray directory on the link line

By consolidating each rpath addition as a single list element even after
whitespace tokenization, it will remain intact.

This was biting us when fetching ipopt.pc, which has several junk
compiler paths thrown into it on Ubuntu 18.04.