tools/performance: Split useful tooling out from cassie_bench
Relevant to: #14464

Gather the sharable bits of experiment execution scripting and port the
cassie benchmark to use it.

This is a first pass at factoring out the common stuff from the
cassie_benchmark tree, for use by other benchmarks. It hoists the basic
techniques for data collection and variance reduction. As this gathers
more uses, it can be polished more.

Also missing from this patch is any documentation or tooling to help
with using the compare.py script from the googlebenchmark package.