tools/performance: Start gathering sharable infrastructure
Relevant to: #14464

Gather the sharable bits of experiment execution scripting and port the cassie
benchmark to use it.

This is a WIP branch, exploring the idea of factoring some experiment-execution
machinery out of the cassie_benchmark example.  I'm not greatly sold on how it
looks so far, though I think it is potentially useful enough and solves enough
distracting subproblems to be worth some sort of better expression.