Refactor FemState and FemElement
Refactor FemState and FemElement to eliminate heap allocation and
virtual methods. Change to a CRTP implemention for compile-time
polymorphism.

Change ElementCacheEntry class to fit it into the new design.

Add an additional attribute, ODE order, to FemElement so that PDEs that
are spatially discretized into ODEs of different orders get a unified
treatment.