Fixes bug in HalfSpace-HalfSpace support for point-pair penetration
The very special case of HalfSpace-HalfSpace was not considered. This
should throw an error as being a meaningless query. Before this patch,
the fcl query would be dispatched and we were at the mercy of how fcl
handled it -- it handles it poorly. Now, we officially report a problem.

This also cleans up the documentationin QueryObject to reflect the current
state.