workspace: Change how fmt and spdlog dependencies are obtained
On Ubuntu 18.04, we downgrade to spdlog 5.1.0 and fmt 6.1.2, but we still
compile them from github source releases.

On Ubuntu 20.04, fmt and spdlog are obtained from the host now.  (The host
copy is also installed by Drake, because Ubuntu's packaging is deficient.)
The versions are spdlog 5.1.0 with bundled fmt 6.1.2.

On macOS, fmt and spdlog are obtained from homebrew now.  The versions are
the latest available, per the usual homebrew policy.