Introduce initial framework for characterizing distance queries
This introduces the CharacterizeResultTest (and supporting
classes). Its purpose is to provide a common framework through
which queries that report signed-distance-related values (e.g.,
point penetration and signed distance) can be evaluated to
determine how good the support is across geometry pairs and scalar
types.

This also includes the *initial* unit test exercising the framework
for the penetration_as_point_pair::Callback. It only evaluates
sphere-sphere contact for T = {double, AutoDiffXd} as a small
example of the framework.

Supporting classes (and functions) in
characterization_utilities.{h|cc}:
 - DistanceCallback - a polymorphic class that each Callback test
   will implement to interface with the framework.
 - Expectation - declaration of what we expect the outcome of the
   query will be.
 - Configuration - a specific configuration between two shapes with
   the known signed distance value between them.
 - to_string() for fcl::NODE_TYPE (to improve error messages).
 - MakeFclShape - a reifier that turns drake::geometry::Shape types
   into the corresponding fcl type.

What follows in subsequent PRs:
  - expand the framework to automatically build configurations in
    a compact, consistent form.
  - complete the tests on penetration_as_point_pair::Callback for
    all other relevant shape-pairs.
  - Modification of documentation in query_object.h reporting the
    results found in test.
  - Extension of framework to evaluate distance_to_shape::Callback
    in the same manner.