[geometry] Improve SceneGraph error for unconnected input ports (#14773)
* Correct SceneGraph error for unconnected input ports

The old error message strictly reported source *id*, which is not
incredibly helpful for the end user. So, now, it includes both
source id and registered name. It further elaborates on the
condition when not connecting is erroneous.

It also removes the lambda function which served purely a
speculative role -- it was waiting for the day when velocities or
accelerations were also available. When we start consuming
velocities, we can restore the generality.