[geometry] Move elements of distance_to_shape.h into .cc (#14772)
* Move elements of distance_to_shape.h into .cc

The distance-to-shape callback was entirely defined in the header
file. This moves as much as possible into the .cc file.

The ComputeNarrowPhaseDistance method is still defined in the
header file to faciltiate the logic for when it invokes the
fallback. The test specializes the fallback function on a unique
scalar (float). The float implementation signals invocation. Then
we call the otherwise undefined CompupteNarrowPhaseDistance<float>
and watch when the special implementation gets triggered.

A note has been added to a test to find an alternate formulation of
this test to detect when the fallback is invoked.